+++ source/Private/Convert-ShouldBeFalse.ps1
@@ -119,19 +119,31 @@ function Convert-ShouldBeFalse
                 If a previous positional parameter is missing then the ones behind
                 it cannot be set to positional.
             #>
-            if ($commandParameters.ActualValue)
+            if ($commandParameters.Because)
             {
-                $commandParameters.ActualValue.Positional = $true
+                $commandParameters.Because.Positional = $true
 
-                if ($commandParameters.Because)
+                if ($commandParameters.ActualValue)
                 {
-                    $commandParameters.Because.Positional = $true
+                    $commandParameters.ActualValue.Positional = $true
                 }
             }
         }
 
         $newExtentText += $commandParameters.ActualValue.Positional ? (' {0}' -f $commandParameters.ActualValue.ExtentText) : ''
-        $newExtentText += $commandParameters.Because.Positional ? (' {0}' -f $commandParameters.Because.ExtentText) : ''
+
+        if ($commandParameters.Because)
+        {
+            # Only add second positional if the first positional was present.
+            if ($commandParameters.ActualValue.Positional)
+            {
+                $newExtentText += $commandParameters.Because.Positional ? (' {0}' -f $commandParameters.Because.ExtentText) : ''
+            }
+            else
+            {
+                $commandParameters.Because.Positional = $false
+            }
+        }
 
         # Holds the new parameter names so they can be added in alphabetical order.
         $parameterNames = @()